Changed LinExpr constructor to take int instead of double (I think double was needed in the old interface for some weird diambiguation reasons, but it works with int now).
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@11941 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/minimodel.hh
@@ -203,7 +203,7 @@ namespace Gecode {
     LinExpr(void);
     /// Create expression for constant \a c
     GECODE_MINIMODEL_EXPORT
-    LinExpr(double c);
+    LinExpr(int c);
     /// Create expression
     GECODE_MINIMODEL_EXPORT
     LinExpr(const IntVar& x, int a=1);
